solana_sanitize

Trait Sanitize

pub trait Sanitize {
    // Provided method
    fn sanitize(&self) -> Result<(), SanitizeError> { ... }
}

A trait for sanitizing values and members of over-the-wire messages.

Implementation should recursively descend through the data structure and sanitize all struct members and enum clauses. Sanitize excludes signature- verification checks, those are handled by another pass. Sanitize checks should include but are not limited to:

All index values are in range.

All values are within their static max/min bounds.

Provided Methods

fn sanitize(&self) -> Result<(), SanitizeError>